fix: move save_to_csv_files into output utils
