+++ graph_datasets/utils/__init__.py
@@ -3,7 +3,6 @@
 # pylint:disable=invalid-name
 from .common import *
 from .evaluation import evaluate_from_embed_file
-from .evaluation import save_to_csv_files
 from .model_management import check_modelfile_exists
 from .model_management import get_modelfile_path
 from .model_management import load_model
@@ -13,6 +12,7 @@ from .model_management import set_seed
 from .output import csv2file
 from .output import make_parent_dirs
 from .output import refresh_file
+from .output import save_to_csv_files
 from .statistics import edge_homo
 from .statistics import node_homo
 from .statistics import statistics
