+++ graph_datasets/utils/evaluation/__init__.py
@@ -1,4 +1,3 @@
 """Evalutaion
 """
 from .evaluation import evaluate_from_embed_file
-from .evaluation import save_to_csv_files
